#58 feat: 일정 그룹 관리 오류 로직 수정중
